+++ src/main/java/com/example/demo/service/ProductoService.java
@@ -6,5 +6,5 @@ import com.example.demo.entity.Producto;
 
 public interface ProductoService {
     Collection<Producto> getAllProductos();
-    Producto getProductoById(int id);
+    Producto getProductoById(Long id);
 }
